+++ src/store/store.hpp
@@ -12,7 +12,7 @@ class Store {
         virtual void init_store(std::string s) = 0;
         virtual void shutdown_store() = 0;
 
-        virtual asio::awaitable<size_t> do_write(std::string o, session_buffer& buffer) = 0;
+        virtual asio::awaitable<int> do_write(std::string o, session_buffer& buffer) = 0;
         virtual asio::awaitable<int> do_read(std::string o, uint64_t offset, session_buffer& buffer) = 0;
         virtual asio::awaitable<bool> do_delete(std::string_view o) = 0;
         virtual asio::awaitable<void> do_list(std::string_view prefix, session_buffer& buffer) = 0;
